refactor: move util function calculate fee pair to helper file
